requirements.txt
uvicorn[standard]~=0.17.5
gunicorn==20.1.0
fastapi~=0.75.0

SQLAlchemy~=1.4.28
pymysql==0.9.3
pydantic~=1.9.0
starlette~=0.17.1
loguru~=0.6.0
alembic~=1.7.5
httplib2~=0.20.2
addict==2.4.0
certifi==2021.10.8
urllib3~=1.26.7
enum34==1.1.10

aiomysql==0.0.22
SQLAlchemy-Utils==0.38.2
python-jose==3.3.0

python-multipart==0.0.5
fastapi-pagination==0.9.1
PyOpenSSL==22.0.0
aiofiles==0.8.0
validators==0.18.2
requests~=2.27.1
redis~=4.2.0

xpinyin~=0.7.6
PyGithub~=1.55
python-gitlab~=2.10.1
PyYAML~=6.0